fastapi_model/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.116.2
numpy==2.3.3
onnxruntime-gpu==1.23.0
opencv-python==4.11.0.86
pika==1.3.2
pillow==11.3.0
pycuda==2025.1.2
pydantic-settings==2.10.1
starlette==0.48.0
tensorrt_cu12==10.13.3.9
uvicorn==0.36.0
